requirements.txt
flask>=3.0
opencv-python-headless>=4.8
kociemba>=1.2
pyserial>=3.5
numpy>=1.24
